Adds wait_for_miq - tails first boot script log
wait_for_miq ssh into MIQ VM and tails log of first boot script injected
via virt-builder. In our case it is output of miq_install script.

The reason why to include this is to let user see what is happening,
that something is happening at all and to immediately see if something
brakes.